Modify ESPTelnet to extend Stream
Defined ESPTelnet to extend Stream in the class definition, which also implicitly extends Print.
Removed custom implementations of print() functions and added implementations for pure virtual functions of Stream and Print.
Followed same paradigm as with original print() functions.  That is, forward them to WiFiClient if and only if there is a valid connected telnet client.