Remove input received callback
To be useful as a Stream object, the client of ESPTelnet should be able to use the built in methods of Stream.  The input received callback precludes this because the ESPTelnet::loop() functions reads out the input before the client would be able to.
The input received callback capability has been removed to make ESPTelnet compatible in any context that accepts a Stream object.